backend/requirements.txt
flask>=3.0.0
pillow>=10.0.0
gunicorn>=21.0.0cryptography>=41.0.0